Maven pom, Eclipse project and a very simple test case.
There's enough here to see what the API should look like, but not enough
to actually *do* anything.  The test case is pretty much just testing
that the classes exist with the right methods in them.  It does run, but
wouldn't do much if you actually tried to pass value through.